Add edit action to track crud

+++ src/Admin/Service/TrackPictureService.php
@@ -23,12 +23,50 @@ final readonly class TrackPictureService
         return is_file($fullPath);
     }
 
+    public function temporaryRename(string $filename): string
+    {
+        $trackPicturesDirectory = $this->parameterBag->get('track_pictures_directory');
+
+        $temporaryFilename = $this->getTemporaryFilename($filename);
+
+        $currentPath = Path::join($trackPicturesDirectory, $filename);
+        $temporaryPath = Path::join($trackPicturesDirectory, $temporaryFilename);
+
+        rename($currentPath, $temporaryPath);
+
+        return $temporaryFilename;
+    }
+
+    public function revertTemporaryRename(string $filename): void
+    {
+        $trackPicturesDirectory = $this->parameterBag->get('track_pictures_directory');
+
+        $temporaryFilename = $this->getTemporaryFilename($filename);
+
+        $temporaryPath = Path::join($trackPicturesDirectory, $temporaryFilename);
+        $originalPath = Path::join($trackPicturesDirectory, $filename);
+
+        rename($temporaryPath, $originalPath);
+    }
+
     public function remove(string $filename): void
     {
         $trackPicturesDirectory = $this->parameterBag->get('track_pictures_directory');
 
         $fullPath = Path::join($trackPicturesDirectory, $filename);
 
+        if (!is_file($fullPath)) {
+            return;
+        }
+
         unlink($fullPath);
+    }
+
+    private function getTemporaryFilename(string $filename): string
+    {
+        $basename = pathinfo($filename, PATHINFO_FILENAME);
+        $extension = pathinfo($filename, PATHINFO_EXTENSION);
+
+        return sprintf('%s_temporary.%s', $basename, $extension);
     }
 }
